Removed MicroservicesMessagingDemo project from the solution and refactored SalesOrderApp with the setting up code in MicroservicesMessagingDemo including the Startup.
