+++ examples/MicroservicesMessagingDemo.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MicroservicesMessagingDemo", "MicroservicesMessagingDemo\MicroservicesMessagingDemo.csproj", "{E96D8131-E13F-44E3-9338-4F61EF80CB96}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SalesOrderApp", "SalesOrderApp\SalesOrderApp.csproj", "{C0E6DEC5-EEAD-4892-A534-93D754B5B993}"
 EndProject
 Global
@@ -17,18 +15,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Debug|x64.Build.0 = Debug|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Debug|x86.Build.0 = Debug|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Release|x64.ActiveCfg = Release|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Release|x64.Build.0 = Release|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Release|x86.ActiveCfg = Release|Any CPU
-		{E96D8131-E13F-44E3-9338-4F61EF80CB96}.Release|x86.Build.0 = Release|Any CPU
 		{C0E6DEC5-EEAD-4892-A534-93D754B5B993}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C0E6DEC5-EEAD-4892-A534-93D754B5B993}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C0E6DEC5-EEAD-4892-A534-93D754B5B993}.Debug|x64.ActiveCfg = Debug|Any CPU
