+++ examples/SalesOrderApp/Startup.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace MicroservicesMessagingDemo
+namespace SalesOrderApp
 {
     public class Startup
     {
