+++ examples/SalesOrderApp/Testing.cs
@@ -1,5 +1,4 @@
 ﻿using System.IO;
-using MicroservicesMessagingDemo;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
